+++ test/rol_adapter_test/brandes_griesse/Weight_Matrices_brandes_griesse.hpp
@@ -81,7 +81,8 @@ public:
       }
      
     HDSA::Ptr<Tpetra::MultiVector<> > z_tpetra = HDSA::dynamicPtrCast<ROL_PDEOPT_Tpetra_Vector<RealT> >(z_in)->get_tpetra_vec(); 
-    Teuchos::Array<long long int> Map_myGlobalIds;
+    //Teuchos::Array<long long int> Map_myGlobalIds;
+    Teuchos::Array<int> Map_myGlobalIds;
     for(int i = 0; i < dim; i++)
       {
 	Map_myGlobalIds.push_back(i);
